WIP repo cleanup
