+++ src/entity.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+
+#[derive(Debug, Deserialize, Serialize, FromRow, PartialEq, Eq)]
+pub struct TodoUser {
+    pub id: i32,
+    pub first_name: String,
+    pub last_name: String,
+}
+
+#[derive(Debug, Deserialize, Serialize, FromRow)]
+pub struct TodoTask {
+    pub id: i32,
+    pub user_id: i32,
+    pub item_desc: String,
+}
